Refactored caching in blog routes module and added a new method to clear cache.
The caching functionality in the `blogRoutes.js` module has been refactored to use a hash key instead of a single key for storing values in Redis. Additionally, a new `clearHash` method has been added to the caching service in `cache.js` to allow clearing of specific hash keys.

This change improves the efficiency and flexibility of caching in the application, allowing for more granular control over cache invalidation.

The cache is now cleared for the user's id when a blog post is created or updated in order to ensure that the most up-to-date data is always served.